third_party/WebKit/LayoutTests/editing/execCommand/insert_list/insert_list_ul_li_to_ol.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable><ul><li>foo</li><li>|bar</li></ul></div>',
    'insertOrderedList',
    '<div contenteditable><ul><li>foo</li></ul><ol><li>|bar</li></ol></div>'),
    'Create OL from last LI of UL');
</script>
